requirements-base.txt
django-configurations==2.4
django-filter==22.1
graphene_django==3.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.2.0
